requirements.txt
fastapi  
uvicorn  
httpx  
websockets
pydantic  
pydantic-settings 
spacy  
sentence-transformers  
gremlinpython  
python-dotenv  
broadcaster[redis]  
redis  